techcorp_support/requirements.txt
crewai>=1.10.1
crewai-tools>=0.25.0

langfuse>=2.50.0
ollama>=0.4.0

streamlit>=1.39.0
pydantic>=2.9.2
pyyaml>=6.0.2
python-dotenv>=1.0.1
rich>=13.9.4
httpx>=0.27.2
tenacity>=9.0.0
langchain-ollama>=0.1.0

